MercatorMidpoint/requirements.txt
flask
google.cloud.aiplatform